// src/app/modules/clients/model/JuniorEducationOrQualificationModel.ts
import { MlcModel } from './common/mlc.model';

export class JuniorEducationOrQualification implements MlcModel {
  name: string;
  location: string;
  datestarted: string;
  datefinished: string;
  schoolidorlogin: string;
  grades: string;
  highlights: string;
  teachernames: string;
  notes: string;
}